Python/requirements.txt
Flask>=2.0
Werkzeug>=2.0
psycopg2-binary>=2.9